Fixed bugs (angle formula, pentagon/heptagon rotation, memory leak with point array)

+++ RG/Lab/LV1-GDI-Tangram/LV1-GDI-Tangram/LV1-GDI-TangramView.h
@@ -14,7 +14,7 @@ private:
 
 // Polygon
 	POINT* drawPolygon(CDC* pDC, POINT* nodes, int nodeNumber, COLORREF color, COLORREF lineColor, int hached);
-	void drawInpolygon(CDC* pDC, POINT* center, int nodeNumber, int radius, double rotation, COLORREF lineColor);
+	void drawInpolygon(CDC* pDC, POINT* center, int nodeNumber, double radius, double rotation, COLORREF lineColor);
 	int distance(POINT& a, POINT& b);
 
 protected: // create from serialization only
